use pub(crate) for elasticsearch internal
and spin out the configuration in a different file so that it is
accessible from the client.
